Refactor checkout skeletons
