--- packages/checkout-storefront/src/sections/Addresses/index.tsx
@@ -1,1 +1,0 @@
-export * from "./AddressesSkeleton";
