+++ packages/checkout-storefront/src/sections/Contact/index.tsx
@@ -1,2 +1,1 @@
 export * from "./Contact";
-export * from "./ContactSkeleton";
