+++ packages/checkout-storefront/src/sections/DeliveryMethods/index.tsx
@@ -1,2 +1,1 @@
 export * from "./DeliveryMethods";
-export * from "./DeliveryMethodsSkeleton";
