+++ packages/checkout-storefront/src/sections/Skeletons/ContactSkeleton.tsx
@@ -5,7 +5,7 @@ interface ContactSkeletonProps {}
 
 export const ContactSkeleton: React.FC<ContactSkeletonProps> = ({}) => {
   return (
-    <div>
+    <div className="section">
       <Skeleton variant="title" />
       <div className="flex flex-row justify-between">
         <Skeleton className="w-1/2" />
