+++ packages/checkout-storefront/src/sections/Summary/index.tsx
@@ -1,3 +1,2 @@
 export * from "./Summary";
 export * from "./FinalizedSummary";
-export * from "./SummarySkeleton";
